# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Schuba API'
  description: 'API for managing diving club activities'
  contact:
    name: 'Groupe 1'
  version: 0.9.0
paths:
...
  /api/group:
    get:
      tags:
        - Groups
      summary: 'Get group records'
      description: 'Retrieves group records based on optional filters for user, initiator, session, and presence status.'
      operationId: 1b29d655f5c291dba95ffff8eadfb856
      parameters:
        -
          name: USER_ID
          in: query
          description: 'User ID to filter group by user'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: INIT_ID
          in: query
          description: 'Initiator ID to filter group by initiator'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: SESS_ID
          in: query
          description: 'Session ID to filter group by session'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: PRESENT
          in: query
          description: 'Presence status of the user in the session'
          required: false
          schema:
            type: boolean
            example: true
      responses:
        200:
          description: 'Successfully retrieved group records'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties:
                    USER_ID:
                      type: integer
                      example: 1
                    INIT_ID:
                      type: integer
                      example: 1
                    SESS_ID:
                      type: integer
                      example: 1
                    PRESENT:
                      type: boolean
                      example: true
                  type: object
        400:
          description: 'Bad request, validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                type: object
        500:
          description: 'Internal server error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something went wrong, please try again.'
                type: object
    post:
      tags:
        - Groups
      summary: 'Create a group for a session'
      description: 'Creates a group for a given session, associating a user with a session and a club. Presence is also recorded.'
      operationId: 4a1769a10757369437f36129be04e2e3
      parameters:
        -
          name: SESS_ID
          in: query
          description: 'Session ID associated with the group'
          required: true
          schema:
            type: integer
            example: 1
        -
          name: USER_ID
          in: query
          description: 'User ID to associate with the group'
          required: true
          schema:
            type: integer
            example: 1
        -
          name: INIT_ID
          in: query
          description: 'Club ID of the initiator for the session'
          required: true
          schema:
            type: integer
            example: 1
        -
          name: PRESENCE
          in: query
          description: 'Presence status of the user in the session'
          required: true
          schema:
            type: boolean
            example: true
      responses:
        200:
          description: 'Group successfully created'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Group successfully created!'
                  session:
                    properties:
                      ID:
                        type: integer
                        example: 1
                      LOCATION_ID:
                        type: integer
                        example: 2
                      LEVEL:
                        type: integer
                        example: 3
                      START:
                        type: string
                        format: date-time
                        example: '2025-01-01T09:00:00'
                      END:
                        type: string
                        format: date-time
                        example: '2025-01-01T11:00:00'
                    type: object
                type: object
        400:
          description: 'Bad request, validation failed'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'The given data was invalid.'
                type: object
        500:
          description: 'Internal server error'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Something went wrong, please try again.'
                type: object
  '/api/group/{id}':
    delete:
      tags:
        - Groups
      summary: 'Delete a Group'
      description: 'Deletes a specific Group based on the provided ID.'
      operationId: 430deed00d0bc83ecb26bb5cde1a67ff
      parameters:
        -
          name: id
          in: path
          description: 'ID of the Group to delete'
          required: true
          schema:
            type: integer
            example: 1
      responses:
        200:
          description: 'Successfully deleted the Group'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Group successfully deleted!'
                type: object
        404:
          description: 'Group record not found'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Group not found.'
                type: object
        500:
          description: 'Internal server error'
...